+++ src/types/customIds.ts
@@ -1,2 +1,3 @@
 export const PRIME_MODAL = 'botluck:prime-modal';
 export const PRIME_MODAL_TEMPLATE = 'template';
+export const PRIME_MODAL_THEME = 'theme';